move ReductionScheduler from registry.cpp to reduction.cpp
